Split navigation into multiple files for cleaner implementation
