+++ src/navigation/trigger.h
@@ -1,0 +1,6 @@
+#ifndef _TRIGGER_H_
+#define _TRIGGER_H_
+
+void Trigger_Setup();
+
+#endif
